requirements.txt
streamlit
opencv-python-headless
numpy
easyocr
Pillow
inference-sdk